BUG Small fixes to analytics quickstart NBs
In notebook 01, I found that the "source" column contained a list; check the "type" field for each element of the list separately.

In notebook 02, the code implied a desired limit to retrieved features based on date, but the cutoff wasn't present. This is the minimal change to add a limit -- it will include features from before the cutoff if they're on the same page as features after the cutoff.

In notebook 03, fix a typo which put the wrong variable name in the `donwload_file` helper function.